re-factoring of deploy module
