+++ docs/development/data_development/data_index.rst
@@ -18,18 +18,18 @@ Different from most other frameworks |daliuge| makes data components first class
 
 Types of Data Components
 ------------------------
-|daliuge| out of the box includes the following Data Components:
+|daliuge| out of the box includes the following lower level Data Components:
 
-#. Posix file data component
-#. Posix directory component
-#. Memory data component
-#. Shared memory data component
-#. S3 data component
-#. NGAS data component
-#. Apache Plasma and Plasma Flight data Components
-#. RDBMS component
+#. `Posix file data component <https://daliuge.readthedocs.io/en/latest/api/dlg.html#dlg.drop.FileDROP>`_
+#. `Posix directory component <https://daliuge.readthedocs.io/en/latest/api/dlg.html#dlg.drop.DirectoryContainer>`_
+#. `Memory data component <https://daliuge.readthedocs.io/en/latest/api/dlg.html#dlg.drop.InMemoryDROP>`_
+#. Shared memory data component (documentation in progress)
+#. `S3 data component <https://daliuge.readthedocs.io/en/latest/api/dlg.html#module-dlg.s3_drop>`_
+#. `NGAS data component <https://daliuge.readthedocs.io/en/latest/api/dlg.html#dlg.drop.NgasDROP>`_
+#. `Apache Plasma <https://daliuge.readthedocs.io/en/latest/api/dlg.html#dlg.drop.PlasmaDROP>`_ and `Plasma Flight <https://daliuge.readthedocs.io/en/latest/api/dlg.html#dlg.drop.PlasmaFlightDROP>`_ data Components
+#. `RDBMS component <https://daliuge.readthedocs.io/en/latest/api/dlg.html#dlg.drop.RDBMSDrop>`_
 
-This range covers most of the use cases in workflows we have encountered sofar, and thus in most cases it is likely not required to develop additional ones. However, adding a new data component is still possible and briefly outlined here. Depending on the I/O model of such a new component, the development can be derived from either one of the existing higher level data components, the lower level :class:`DataIO <dlg.io.DataIO>` abstract class, or even lower with the :class:`AbstractDROP <dlg.drop.AbstractDROP>` class. The best way to get started is to checkout the code of existing data components like , e.g. the :class:`JsonDROP <dlg.json_drop.JsonDROP>` or the the :class:`S3DROP <dlg.s3_drop.S3DROP>`.
+This range covers most of the use cases in workflows we have encountered sofar, and thus in most cases it is likely not required to develop additional ones. However, adding a new data component is still possible and briefly outlined here. Depending on the I/O model of such a new component, the development can be derived from either one of the existing higher level data components, the lower level :class:`DataIO <dlg.io.DataIO>` abstract class, or even lower with the :class:`AbstractDROP <dlg.drop.AbstractDROP>` class. The best way to get started is to checkout the code of existing data components like , e.g. the very high level :class:`JsonDROP <dlg.json_drop.JsonDROP>` or the fairly complex, lower level :class:`S3DROP <dlg.s3_drop.S3DROP>`.
 
 I/O
 ---
